Gerber Files
I recommend ordering these from Aisler.net
a Maker run business from Germany/Netherlands. As of ordering this Board myself the quote including shipping(to Austria) was the same as ordering these boards from China, but it was a bit faster. And they have a great personal support.
+ you can order it as a Kit including Parts from Aisler.net (the Author gets a Kickback  when you order a Kit)
All the parts needed are available at Farnell and Newark